+++ src/components/icons/ManagementClusterIcon.tsx
