+++ src/components/icons/ProtectedClusterIcon.tsx
