+++ src/icons.ts
@@ -174,9 +174,9 @@ export { default as WindowsLogoIcon } from './components/icons/WindowsLogoIcon'
 export { default as WorkspaceIcon } from './components/icons/WorkspaceIcon'
 export { default as InfoOutlineIcon } from './components/icons/InfoOutlineIcon'
 export { default as PipelineIcon } from './components/icons/PipelineIcon'
-export { default as ProtectedClusterIcon } from './components/icons/ProtectedCluster'
-export { default as ManagementClusterIcon } from './components/icons/ManagementCluster'
-export { default as ShieldOutlineIcon } from './components/icons/ShieldOutline'
+export { default as ProtectedClusterIcon } from './components/icons/ProtectedClusterIcon'
+export { default as ManagementClusterIcon } from './components/icons/ManagementClusterIcon'
+export { default as ShieldOutlineIcon } from './components/icons/ShieldOutlineIcon'
 // Logos
 export { default as PluralLogoFull } from './components/icons/logo/PluralLogoFull'
 export { default as PluralLogoMark } from './components/icons/logo/PluralLogoMark'
